### 2025-12-23 #318 @editorialbot
[PRE REVIEW]: scicode-widgets: Bringing Computational Experiments to the Classroom with Jupyter Widgets
**Submitting author:** @agoscinski (<a href="http://orcid.org/0000-0001-8076-215X">Alexander Goscinski</a>)
**Repository:** https://github.com/osscar-org/scicode-widgets
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.2.0
