### 2025-12-29 #318 @arm61
hey @agoscinski, I was wondering about the difference between the committer and the authors on the paper. Could you please comment on the role of those not listed among the commiters that are listed as co-author?
