### 2025-12-29 #318 @agoscinski
This is because we initially developed the package in this repo https://github.com/ceriottm/iam-notebooks where the code was bundled with the course material. Later on, the package has been moved into its own repo referenced in the paper. The people who initially contributed already left so they are not listed as contributor in the new repo.
