### 2026-01-08 #318 @agoscinski
Thank you for the list. Here is a list of people we identified as good candidates for the review. The last two people are not from the list so I added their contact information.
- Tania Allard
- Matthias Bussonnier
- Bryan Weber
- Daniel Hollas [Github danielhollas](https://github.com/danielhollas) [Contact page](https://www.bristol.ac.uk/people/person/Daniel-Hollas-e03e6bd9-7e6e-48a2-98ac-dd4746d820b7)
- Chris Holdgraf [Github choldgraf](https://github.com/choldgraf) [Contact page](https://www.icsi.berkeley.edu/icsi/people/holdgraf)

Full disclosure, we collaborate with Daniel Hollas on a completely different software project that is not related to this paper in any way.
