### 2026-01-21 #318 @arm61
@trallard or @choldgraf would either of you be interested to review this JOSE work?
