### 2026-01-21 #318 @choldgraf
I don't believe I'm qualified to review a paper about widgets. I haven't done software engineering on that part of the jupyter stack much, and certainly not in the last few years. Sorry! Maybe @jasongrout would be interested if he has the time?
