### 2026-01-21 #318 @jasongrout
Sure, I'm interested. Before committing, though, what expectations do you have and what information do you want from a reviewer? Is there a form the reviewer fills out?
