### 2026-01-21 #318 @jasongrout
https://jose.theoj.org/about links to [this review process](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html). Did I find the correct set of expectations for reviewers?
