### 2026-01-29 #318 @Carreau
I'm a bit overwhelmed at the time, I would prefer to decline for the time being. But thanks for thinking about me, and I'll add it to my long to-read list of things :-)
