### 2026-02-02 #318 @nvaytet
Flattered that you thought of me @arm61 , and the article looks very interesting, but I will simply not have the time to do this at the moment 😞
Nice to see that you got further involved in the open journal, though!
