### 2026-02-14 9468f9a
Fix a bug where the check registry is not correctly initialized
