### 2026-03-20 #326 @arm61
@jasongrout and @jd15489, thanks again for agreeing to review. If you work through the checklist (information above on how to generate it) and if there are any problems/comments about the material, I would recommend opening an issue on [the material repository](https://github.com/osscar-org/scicode-widgets), and the authors can sort them out.

More information about the review guidelines can be found on the Open Journals documentation pages: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html

If anyone has any questions just ping me on here!
