### 2026-03-20 #326 @arm61
@agoscinski can you also address the missing section points highlighted above?
