+++ paper.bib
@@ -94,6 +94,8 @@
 	pages = {108546},
 	title = {{OSSCAR}, an open platform for collaborative development of computational tools for education in science},
 	volume = {282},
+	url = {https://doi.org/10.1016/j.cpc.2022.108546},
+	DOI = {10.1016/j.cpc.2022.108546},
 	year = {2023}}
 
 
@@ -103,6 +105,8 @@
 	pages = {109353},
 	title = {Jupyter widgets and extensions for education and research in computational physics and chemistry},
 	volume = {305},
+	url = {https://doi.org/10.1016/j.cpc.2024.109353},
+	DOI = {10.1016/j.cpc.2024.109353},
 	year = {2024}}
 
 
@@ -196,8 +200,7 @@ note = {Accessed 4 July 2025}
     year = {2006},
     isbn = {0387310738},
     publisher = {Springer-Verlag},
-    address = {Berlin, Heidelberg},
-    DOI = {10.5555/1162264}
+    address = {Berlin, Heidelberg}
 }
 
 @misc{selenium,
