### 2026-03-25 #326 @jd15489
- [x] **Version:** Does the release version given match the GitHub release?
