+++ CONTRIBUTING.rst
@@ -1,4 +1,4 @@
 Contributing
 ============
 
-Please read the `developer documentation <developers.html>`_.
+Please read the `developer documentation <https://scicode-widgets.readthedocs.io/en/latest/developers.html>`_.
