### 2026-03-30 0bf2214
add a character count mechanism for textexercise
Limit too verbose answers
