### 2026-07-02 168d1ec
Update CONTRIBUTING.rst with issue reporting details
Added instructions for reporting issues and support contacts.
