### 2026-07-02 0f22c7e
Add issue reporting and support request sections to README
