### 2026-07-02 7e34522
Update CONTRIBUTING.rst
Removed contact information for support requests.
