+++ CONTRIBUTING.rst
@@ -1,11 +1,4 @@
 Contributing
 ============
 
-Please read the `developer documentation <https://scicode-widgets.readthedocs.io/en/latest/developers.html>`_.
-If you encounter any issues with the code, please report them by opening a
-`GitHub issue <https://github.com/osscar-org/scicode-widgets/issues/new>`_. For
-support requests, please contact any of the people listed below: 
-
-- `Alexander Goscinski <https://www.psi.ch/de/lms/people/alexander-jan-goscinski>`_
-- `Giovanni Pizzi <https://www.psi.ch/en/lms/people/giovanni-pizzi>`_
-- `Michele Ceriotti <https://people.epfl.ch/michele.ceriotti?lang=en>`_
+For contributions to the code, please refer to the `developer documentation <https://scicode-widgets.readthedocs.io/en/latest/developers.html>`_.
